trivial: move alerting hooks to a subfolder
risk: low
