+++ libs/sdk-ui-dashboard/src/presentation/widget/insight/configuration/InsightAlertConfig/hooks/useUpdateAlert.ts
@@ -6,7 +6,7 @@ import {
     CommandProcessingStatus,
     saveAlert,
     useDashboardCommandProcessing,
-} from "../../../../../model/index.js";
+} from "../../../../../../model/index.js";
 
 /**
  * @internal
